feat(DelimiterParser): DelimiterParser로 구분 및 StringAddCalculator 수정 반영
